+++ src/generator/mod.rs
@@ -1,7 +1,7 @@
 mod from_proto;
 mod generator;
-mod name;
+mod graphql_type;
 mod source;
 pub use generator::Generator;
-pub use name::GraphQLType;
+pub use graphql_type::GraphQLType;
 pub use source::Source;
